+++ CMakeLists.txt
@@ -52,6 +52,7 @@ endif()
 # Create target and set properties
 add_library(${PROJECT_NAME}
     src/ic_app.cpp
+    src/ic_camera.cpp
     src/ic_log.cpp
     src/ic_graphics.cpp
 )
