+++ include/ic.h
@@ -3,6 +3,7 @@
 #include "ic_camera.h"
 #include "ic_graphics.h"
 #include "ic_log.h"
+#include "ic_renderer.h"
 
 #if ICENGINE_EXPORTS
 __declspec(dllexport)
