+++ include/ic_app.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <renderer.h>
+#include <ic_renderer.h>
 
 namespace IC
 {
