+++ src/vulkan/descriptors.cpp
@@ -1,6 +1,5 @@
-#include <descriptors.h>
-
-#include <swap_chain.h>
+#include "descriptors.h"
+#include "swap_chain.h"
 
 namespace IC::Renderer
 {
