+++ src/vulkan/vulkan_device.cpp
@@ -1,6 +1,5 @@
-#include <vulkan_device.h>
-
-#include <vulkan_types.h>
+#include "vulkan_device.h"
+#include "vulkan_types.h"
 
 // std headers
 #include <cstring>
