+++ src/vulkan/vulkan_util.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <vulkan_types.h>
-#include <vulkan_device.h>
+#include "vulkan_types.h"
+#include "vulkan_device.h"
 
 namespace IC::Renderer::Util
 {
